backend/tests/test_heic_support.py
#!/usr/bin/env python3
"""
Test script for HEIC file support

Tests that .heic and .heif files can be uploaded and processed
"""

import sys
import os
sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

from PIL import Image
from pillow_heif import register_heif_opener

# Register HEIF opener
register_heif_opener()

def test_heic_format_detection():
    """Test that HEIC format is properly detected"""
    print("Testing HEIC format detection...")

    # Test that pillow-heif is properly registered
    try:
        # This will work if pillow-heif is properly installed and registered
        print("✅ pillow-heif library is installed and registered")
        print(f"   Supported formats: {Image.registered_extensions()}")

        # Check if .heic and .heif are in registered extensions
        extensions = Image.registered_extensions()
        if '.heic' in extensions or '.heif' in extensions:
            print("✅ HEIC/HEIF extensions are registered with Pillow")
        else:
            print("⚠️  HEIC/HEIF extensions not found in registered formats")

        return True
    except Exception as e:
        print(f"❌ Error: {e}")
        return False


def test_allowed_extensions():
    """Test that HEIC is in allowed extensions"""
    print("\nTesting allowed extensions configuration...")

    from app import ALLOWED_EXTENSIONS

    if 'heic' in ALLOWED_EXTENSIONS and 'heif' in ALLOWED_EXTENSIONS:
        print(f"✅ HEIC/HEIF are in ALLOWED_EXTENSIONS: {ALLOWED_EXTENSIONS}")
        return True
    else:
        print(f"❌ HEIC/HEIF not in ALLOWED_EXTENSIONS: {ALLOWED_EXTENSIONS}")
        return False


def test_file_upload_endpoint():
    """Test that the API endpoint accepts HEIC files"""
    print("\nTesting API endpoint with HEIC file type...")

    import requests
    from io import BytesIO

    try:
        # Create a dummy HEIC file for testing (just checking if endpoint accepts the extension)
        # We'll use a small test image
        test_image = Image.new('RGB', (100, 100), color='red')
        buffer = BytesIO()
        test_image.save(buffer, format='JPEG')
        buffer.seek(0)

        # Test with .heic extension
        files = {'photo': ('test.heic', buffer, 'image/heic')}

        # Just check if the endpoint doesn't reject the file type
        # (It will fail validation because it's not a real HEIC, but that's OK)
        response = requests.post('http://localhost:5001/api/analyze', files=files, timeout=10)

        # We expect either 200 (if it processes) or 400 with a specific error
        # But NOT "Invalid file type" error
        if response.status_code == 400:
            error = response.json().get('error', '')
            if 'Invalid file type' in error:
                print(f"❌ Endpoint rejected HEIC file type: {error}")
                return False
            else:
                print(f"✅ Endpoint accepts HEIC file type (validation error is normal for test data)")
                return True
        elif response.status_code == 200:
            print("✅ Endpoint successfully processed test file")
            return True
        else:
            print(f"⚠️  Unexpected response: {response.status_code}")
            return True  # Still counts as not rejecting the file type

    except Exception as e:
        print(f"❌ Error testing endpoint: {e}")
        return False


def main():
    """Run all HEIC support tests"""
    print("🧪 StoryBeats HEIC Support Test Suite")
    print("=" * 80)

    results = []

    # Test 1: Format detection
    results.append(("Format Detection", test_heic_format_detection()))

    # Test 2: Allowed extensions
    results.append(("Allowed Extensions", test_allowed_extensions()))

    # Test 3: API endpoint
    results.append(("API Endpoint", test_file_upload_endpoint()))

    # Summary
    print("\n" + "=" * 80)
    print("📊 Test Summary")
    print("=" * 80)

    passed = sum(1 for _, result in results if result)
    total = len(results)

    for test_name, result in results:
        status = "✅ PASSED" if result else "❌ FAILED"
        print(f"  {status}: {test_name}")

    print(f"\n{passed}/{total} tests passed")

    if passed == total:
        print("🎉 All HEIC support tests passed!")
        print("\nYou can now upload iPhone .heic photos to StoryBeats!")
    else:
        print("⚠️  Some tests failed. Check output above.")

    print("=" * 80 + "\n")


if __name__ == "__main__":
    main()
