backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
spotipy==2.23.0
pillow==10.4.0
pillow-heif==0.16.0
anthropic==0.18.1
openai==1.12.0
httpx==0.27.2
requests==2.31.0
python-multipart==0.0.9
werkzeug==3.0.1
gunicorn==21.2.0
